layout/base/tests/test_bug518777.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=518777
-->
<head>
  <title>Test for Bug 518777</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script>
  SimpleTest.waitForExplicitFinish();

  function dotest() {
    var canvasWidth = 50;
    var canvasHeight = 50;
    var source = document.getElementById("source").contentWindow;
    rect = { left: 25, top: 25,
             width: canvasWidth, height: canvasHeight };
    var canvas = SpecialPowers.snapshotRect(source, rect, "transparent");
    var context = canvas.getContext("2d");

    var data = context.getImageData(0, 0, canvasWidth, canvasHeight).data;
    var i;
    for (i = 0; i < data.length; i += 4) {
      if (data[i] != 0 || data[i + 1] != 0 || data[i + 2] != 255 || data[i + 3] != 255)
        break;
    }
    ok(i >= data.length, "all pixels blue");

    SimpleTest.finish();
  }
  </script>
</head>
<body>
<p><a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=518777">Mozilla Bug 518777</a></p>

<iframe id="source" width="50" height="50"
       srcdoc="<html><body onload='window.scrollTo(0,99999999); document.documentElement.offsetWidth; window.parent.dotest();' style='background: rgb(0,0,255); width: 100px; height: 50100px;'></body></html>"></iframe>

<pre id="test">
</pre>
</body>
</html>
